Update now playing by removing unused features and added improvements of easy song controls, and checking for dependencies, code usages, and all quality aspects like securities, etc. And finally, update the latest description / information about the app :)

+++ app/src/main/java/com/musicvault/data/db/SongDao.kt
@@ -44,7 +44,7 @@ interface SongDao {
     fun getDistinctFolders(): LiveData<List<FolderInfo>>
 
     @Query("UPDATE songs SET pitchSemitones = :pitch WHERE id = :id")
-    suspend fun updatePitch(id: Long, pitch: Int)
+    suspend fun updatePitch(id: Long, pitch: Float)
 
     @Query("UPDATE songs SET playbackSpeed = :speed WHERE id = :id")
     suspend fun updateSpeed(id: Long, speed: Float)
